requirements.txt
dash==3.4.0
pymongo==4.16.0
python-dotenv==1.2.1
gunicorn==23.0.0
certifi